requirements.txt
# Core
fastapi
uvicorn[standard]
python-multipart
requests
httpx
pydantic
python-dotenv

# Image processing
pillow
opencv-python-headless
numpy
ImageHash
PyWavelets

# OCR
easyocr

# AI / LLM
groq
tenacity>=8.2.0
duckduckgo-search

# Supabase (production database)
supabase

# Rate limiting
slowapi

# PDF export
reportlab

# WhatsApp
twilio

# Payments
razorpay

# Testing
pytest
pytest-asyncio
